client/python/requirements.txt
grpcio-tools
colorama
